+++ src/node-transformers/converting-transformers/SplitStringTransformer.ts
@@ -23,6 +23,8 @@ import { NodeUtils } from '../../node/NodeUtils';
  */
 @injectable()
 export class SplitStringTransformer extends AbstractNodeTransformer {
+    private static readonly maxStringLengthForSecondPass: number = 30000;
+
     /**
      * @type {number}
      */
@@ -87,7 +89,7 @@ export class SplitStringTransformer extends AbstractNodeTransformer {
         switch (nodeTransformationStage) {
             case NodeTransformationStage.Converting:
                 return {
-                    enter: (node: ESTree.Node, parentNode: ESTree.Node | null): ESTree.Node | undefined => {
+                    leave: (node: ESTree.Node, parentNode: ESTree.Node | null): ESTree.Node | undefined => {
                         if (parentNode && NodeGuards.isLiteralNode(node)) {
                             return this.transformNode(node, parentNode);
                         }
@@ -113,23 +115,34 @@ export class SplitStringTransformer extends AbstractNodeTransformer {
         }
 
         // pass #1: split string on a large chunks with length of `firstPassChunkLength`
-        const firstPassChunksNode: ESTree.Node = this.transformLiteralNodeByChunkLength(
+        const [firstPassChunksNode, stringChunks]: [ESTree.Node, string[]] = this.transformLiteralNodeByChunkLength(
             literalNode,
             SplitStringTransformer.firstPassChunkLength
         );
 
-        // pass #2: split large chunks on a chunks with length of `splitStringsChunkLength`
+        const stringChunksCount: number = stringChunks.length;
+        const isLargeString: boolean = SplitStringTransformer.maxStringLengthForSecondPass
+            / SplitStringTransformer.firstPassChunkLength
+            <= stringChunksCount;
+        const minSecondPathChunkLength = isLargeString
+            ? Math.floor(stringChunksCount / this.options.splitStringsChunkLength)
+            : this.options.splitStringsChunkLength;
+
+        // pass #2: split large chunks on a chunks with length of min length === `splitStringsChunkLength`
         const secondPassChunksNode: ESTree.Node = estraverse.replace(firstPassChunksNode, {
             // eslint-disable-next-line @typescript-eslint/no-shadow
-            enter: (node: ESTree.Node, parentNode: ESTree.Node | null) => {
+            enter: (node: ESTree.Node) => {
                 if (NodeGuards.isLiteralNode(node)) {
                     return this.transformLiteralNodeByChunkLength(
                         node,
-                        this.options.splitStringsChunkLength
-                    );
+                        minSecondPathChunkLength
+                    )[0];
                 }
             }
         });
+
+        NodeUtils.parentizeNode(secondPassChunksNode, parentNode);
+        NodeUtils.parentizeAst(secondPassChunksNode);
 
         NodeUtils.parentizeNode(secondPassChunksNode, parentNode);
         NodeUtils.parentizeAst(secondPassChunksNode);
@@ -140,20 +153,20 @@ export class SplitStringTransformer extends AbstractNodeTransformer {
     /**
      * @param {Literal} literalNode
      * @param {number} chunkLength
-     * @returns {Node}
+     * @returns {[resultNode: Node, stringChunks: string[]]}
      */
     private transformLiteralNodeByChunkLength (
         literalNode: ESTree.Literal,
         chunkLength: number
-    ): ESTree.Node {
+    ): [resultNode: ESTree.Node, stringChunks: string[]] {
         if (!NodeLiteralUtils.isStringLiteralNode(literalNode)) {
-            return literalNode;
+            return [literalNode, []];
         }
 
         const valueLength: number = stringz.length(literalNode.value);
 
         if (chunkLength >= valueLength) {
-            return literalNode;
+            return [literalNode, []];
         }
 
         const stringChunks: string[] = SplitStringTransformer.chunkString(
@@ -162,7 +175,10 @@ export class SplitStringTransformer extends AbstractNodeTransformer {
             chunkLength
         );
 
-        return this.transformStringChunksToBinaryExpressionNode(stringChunks);
+        return [
+            this.transformStringChunksToBinaryExpressionNode(stringChunks),
+            stringChunks
+        ];
     }
 
     /**